[TASK] Restructure new files
